Fix: bulletproof math against zero-dimension thumbnails and unsigned underflows
A relentless final code audit identified and resolved two theoretical edge-cases that could cause erratic behavior or crashes on specific architectures or corrupted media:

- menu-image.c: Guarded the thumbnail cache's aspect ratio calculation against division-by-zero exceptions (FPE) if a corrupted image file loads with 0x0 pixel dimensions.

- control-image.c: Added explicit (int) casts to the zoom window calculations. This prevents potentially dangerous unsigned underflow wrapping when subtracting OSD width from the source image dimensions, guaranteeing safe negative numbers on all processor architectures.

+++ control-image.c
@@ -789,8 +789,8 @@ void cImageControl::ConvertZoom()
   m_ePlayMode = ePlayModeZoom;
 
   // How may pixel are outside screen after zoom
-  m_nZoomXMax = ((m_nRealImageWidth  * m_nZoomFactor) - nMaxWidth);
-  m_nZoomYMax = ((m_nRealImageHeight * m_nZoomFactor) - nMaxHeight);
+  m_nZoomXMax = (int)(m_nRealImageWidth  * m_nZoomFactor) - (int)nMaxWidth;
+  m_nZoomYMax = (int)(m_nRealImageHeight * m_nZoomFactor) - (int)nMaxHeight;
 
   // If image bigger than screen, how many step can i'm move in zoomed image
   if(m_nZoomXMax > 0)
